changed tar test with and without report
